refactor: 공용으로 사용할 수 있도록 input 컴포넌트 분리
