+++ app/mypage/_components/UserProfile/UserProfile.tsx
@@ -13,7 +13,6 @@ import UpdateProfileModal from '../UpdateProfileModal/UpdateProfileModal';
 
 const cx = classNames.bind(styles);
 
-// TODO: 타입 정의
 export default function UserProfile({ userId }: { userId: number }) {
   const { toggleModal, name, setCurrentName } = useModalStore();
