+++ app/signin/_components/RegisterModal/RegisterModal.module.scss
@@ -1,15 +1,13 @@
 @import '/styles/global-styles/index';
-
+// TODO: 모바일 반응형 추가
 .form {
   display: flex;
   width: 39.5rem;
   padding: 3.2rem 2.8rem 3.2rem 3.6rem;
   flex-direction: column;
   align-items: flex-start;
-  //   gap: 48rem;
   background: $BDS-Grayscale-1;
 
-  /* BDS Effect/Writing Modal */
   box-shadow: 0px 4px 15px 0px rgba(0, 0, 0, 0.5);
 }
 
@@ -28,7 +26,7 @@
   font-size: 2rem;
   font-style: normal;
   font-weight: 600;
-  line-height: 140%; /* 28px */
+  line-height: 140%;
 }
 
 .close {
@@ -42,13 +40,14 @@
   display: flex;
   flex-direction: column;
   justify-content: flex-start;
-  margin-bottom: 3.2rem;
+  gap: 3.2rem;
 }
 
 .buttons {
   display: flex;
-  // justify-content: space-between;
   align-items: center;
   justify-content: flex-end;
   gap: 1rem;
+  margin-top: 3.2rem;
+  margin-left: auto;
 }
