+++ app/signin/_components/SignInput/SignInput.module.scss
@@ -1,0 +1,28 @@
+@import '/styles/global-styles/index';
+
+.input-and-check {
+  display: flex;
+  flex-direction: column;
+  align-items: flex-start;
+  gap: 0.4rem;
+}
+
+.message {
+  text-align: center;
+
+  font-size: 1.2rem;
+
+  font-weight: 400;
+  line-height: 133%; /* 15.96px */
+}
+
+.error-text {
+  color: $BDS-System-Error;
+}
+// .error-input {
+//   border-color: $BDS-System-Error;
+// }
+.success {
+  color: $BDS-System-Success;
+  border-color: $BDS-System-Success;
+}
