+++ components/Button/BlobButton.module.scss
@@ -11,7 +11,7 @@
   font-weight: 400;
   line-height: 146%; /* 1.36875rem */
 
-  width: 4.375rem;
+  // width: 4.375rem;
   height: 2.375rem;
   display: inline-flex;
   padding: 0.5rem 1rem;
@@ -46,7 +46,7 @@
   background-clip: text;
   -webkit-background-clip: text;
 
-  width: 4.375rem;
+  // width: 4.375rem;
   height: 2.375rem;
   display: inline-flex;
   padding: 0.5rem 1rem;
