+++ components/Input/Input.module.scss
@@ -10,7 +10,7 @@
   @include flex-column;
 
   gap: 1.2rem;
-  margin-bottom: 3.2rem;
+  // margin-bottom: 3.2rem;
 }
 
 .label {
